update tests, models and salt generator

+++ Models/Transaction.cs
@@ -4,7 +4,6 @@
     {
         public string Sender { get; set; }
         public string Receiver { get; set; }
-        public int Amount { get; set; }
-
+        public double Amount { get; set; }
     }
 }
